Move AlertDestination to correct PSR4 namespace
